+++ internal/analyzer/analyzer_test.go
